Add files via upload
Exercise 3 
Create a Node Prompt that asks a question
Then have the prompt print to the console a string with some of the user's response
Then after that question is asked. Ask another question
Put the input logic and the output logic into different files and require both of them into a main file. Make the main file callable from the cli.